+++ libs/admin/src/lib/admin-panel/pages/overview/overview.component.scss
@@ -1,3 +1,7 @@
 section {
   margin: 40px;
 }
+
+li a mat-icon {
+  vertical-align: middle;
+}
